move channelManager channelState to channel_manager.rs
