+++ miner-apps/translator/src/lib/sv2/channel_manager/mod.rs
@@ -2,4 +2,3 @@ pub mod channel_manager;
 pub mod extensions_message_handler;
 pub mod mining_message_handler;
 pub use channel_manager::ChannelManager;
-pub(super) mod channel;
